# Improve Drawrule matching performance