--- a/core/src/data/dataSource.cpp
+++ b/core/src/data/dataSource.cpp
@@ -12,6 +12,7 @@
 #include <mutex>
 #include <list>
 #include <functional>
+#include <unordered_map>
 
 namespace Tangram {