--- a/core/src/scene/drawRule.h
+++ b/core/src/scene/drawRule.h
@@ -2,33 +2,39 @@
 
 #include "scene/styleParam.h"
 
-#include <string>
 #include <vector>
 
 namespace Tangram {
 
+class Style;
+class Scene;
+class Tile;
 class StyleContext;
+struct Feature;
+struct StaticDrawRule;
 
 struct DrawRule {
 
+    // Reference to original StyleParams
+    const StyleParam* params[StyleParamKeySize];
+
+    // Evaluated params for stops and functions
+    StyleParam evaluated[StyleParamKeySize];
+
     std::string styleName;
     int styleId;
-    std::vector<StyleParam> parameters;
 
-    DrawRule(std::string _styleName, int _styleId, const std::vector<StyleParam>& _parameters,
-             bool _sorted = false);
-
-    //Merge properties of _other and retain self
-    DrawRule merge(DrawRule& _other) const;
-    std::string toString() const;
-
-    bool eval(const StyleContext& _ctx);
+    bool isJSFunction(StyleParamKey _key) const {
+        auto& param = findParameter(_key);
+        if (!param) {
+            return false;
+        }
+        return param.function >= 0;
+    }
 
     const std::string& getStyleName() const;
 
     const StyleParam& findParameter(StyleParamKey _key) const;
-
-    bool isJSFunction(StyleParamKey _key) const;
 
     template<typename T>
     bool get(StyleParamKey _key, T& _value) const {
@@ -40,14 +46,22 @@ struct DrawRule {
         _value = param.value.get<T>();
         return true;
     }
+
     bool contains(StyleParamKey _key) const {
         return findParameter(_key) != false;
     }
 
-    bool operator<(const DrawRule& _rhs) const;
-
 private:
-    void logGetError(StyleParamKey _expectedKey, const StyleParam& _param);
+    void logGetError(StyleParamKey _expectedKey, const StyleParam& _param) const;
+
+};
+
+struct Styling {
+    std::vector<DrawRule> styles;
+
+    void apply(Tile& _tile, const Feature& _feature, const Scene& _scene, StyleContext& _ctx);
+
+    void mergeRules(const std::vector<StaticDrawRule>& rules);
 };
 
 }