--- a/core/src/scene/sceneLoader.h
+++ b/core/src/scene/sceneLoader.h
@@ -24,7 +24,7 @@ class ShaderProgram;
 class Material;
 class Style;
 struct StyleParam;
-struct DrawRule;
+struct StaticDrawRule;
 struct MaterialTexture;
 struct Filter;