… bench/src/tileLoading.cpp
… core/src/scene/scene.cpp
… core/src/scene/scene.h
… core/src/scene/sceneLayer.cpp
… core/src/scene/sceneLayer.h
… core/src/scene/sceneLoader.cpp
… core/src/scene/styleContext.cpp
… core/src/scene/styleContext.h
… core/src/scene/styleParam.cpp
… core/src/scene/styleParam.h
… core/src/style/style.h
… core/src/tile/tile.cpp
… core/src/tile/tile.h
… core/src/tile/tileWorker.cpp
… core/src/util/fastmap.h
… tests/unit/drawRuleTests.cpp
… tests/unit/dukTests.cpp
… tests/unit/labelsTests.cpp
… tests/unit/layerTests.cpp
… toolchains/linux.cmake
… travis/script_run_bench.sh