@hjanetzek 2015-11-13
Some tile build times for 10/301/384.mvt (on a xeon 2560). This is a pretty heavy tile, about 750kb:
For eraser-map style build time (style matching and geometry creation, without parsing) went down from 110ms to 65ms. For the default style from 68ms to 45ms.

Unfortunately one cannot build google benchmark on android so I will try it there with simple chrono counters.

```
Benchmark                             Time(ns)    CPU(ns) Iterations

# base: eraser
TileLoadingFixture/BuildTest_mean    110602791  110466660          5
TileLoadingFixture/BuildTest_stddev    1931617    2001095          0

# perf: eraser
TileLoadingFixture/BuildTest_mean     73978121   73952386          7
TileLoadingFixture/BuildTest_stddev    2108898    2108727          0

# perf: eraser - property key size as primary sort key
TileLoadingFixture/BuildTest_mean     65133547   65083325          8
TileLoadingFixture/BuildTest_stddev    2477973    2491654          0


# base: default
TileLoadingFixture/BuildTest_mean     67846241   67833338          8
TileLoadingFixture/BuildTest_stddev    3576580    3650531          0

# perf: default
TileLoadingFixture/BuildTest_mean     48632142   48027775         12
TileLoadingFixture/BuildTest_stddev    2522254    2348654          0

# perf: default - property key size as primary sort key
TileLoadingFixture/BuildTest_mean     44576521   44564100         13
TileLoadingFixture/BuildTest_stddev    2983526    2947713          0
```