@hjanetzek 2015-11-24
:+1: for tackling the hard problem of variable naming :)