@blair1618 2015-11-24
:+1: for fixing syntax in those tests